put some of the same codes into one class and combine all three validators into one file
